scripts/requirements.txt
torch
torchvision
torchaudio
torchmetrics
tqdm
numpy
pandas
matplotlib
seaborn
scikit-learn
Pillow
opencv-python
Augmentor
fastapi
uvicorn
